requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
